+++ Sources/TurboNavigator/TurboNavigator.swift
@@ -1,3 +1,4 @@
+import SafariServices
 import Turbo
 import UIKit
 import WebKit
@@ -5,18 +6,31 @@ import WebKit
 /// Implement to be notified when certain navigations are performed
 /// or to render a native controller instead of a Turbo web visit.
 public protocol TurboNavigationDelegate: AnyObject {
-    /// Implement to override or customize the controller to be displayed.
-    /// Return `nil` to not display or route anything.
-    func controller(_ controller: VisitableViewController, forProposal proposal: VisitProposal) -> UIViewController?
-
     /// An error occurred loading the request, present it to the user.
     /// Retry the request by calling `session.reload()`.
     func session(_ session: Session, didFailRequestForVisitable visitable: Visitable, error: Error)
+
+    /// Optional. Implement to override or customize the controller to be displayed.
+    /// Return `nil` to not display or route anything.
+    func controller(_ controller: VisitableViewController, forProposal proposal: VisitProposal) -> UIViewController?
+
+    /// Optional. Implement to customize handling of external URLs.
+    /// If not implemented, will present `SFSafariViewController` as a modal and load the URL.
+    func openExternalURL(_ url: URL, from controller: UIViewController)
 }
 
 public extension TurboNavigationDelegate {
     func controller(_ controller: VisitableViewController, forProposal proposal: VisitProposal) -> UIViewController? {
         VisitableViewController(url: proposal.url)
+    }
+
+    func openExternalURL(_ url: URL, from controller: UIViewController) {
+        let safariViewController = SFSafariViewController(url: url)
+        safariViewController.modalPresentationStyle = .pageSheet
+        if #available(iOS 15.0, *) {
+            safariViewController.preferredControlTintColor = .tintColor
+        }
+        controller.present(safariViewController, animated: true)
     }
 }
 
@@ -213,6 +227,11 @@ extension TurboNavigator: SessionDelegate {
         if session == modalSession {
             self.session.clearSnapshotCache()
         }
+    }
+
+    public func session(_ session: Session, openExternalURL url: URL) {
+        let controller = session === modalSession ? modalNavigationController : navigationController
+        delegate?.openExternalURL(url, from: controller)
     }
 
     public func session(_ session: Session, didFailRequestForVisitable visitable: Visitable, error: Error) {